+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Head
 
 - Added: `initialFocus` and `fallbackFocus` options can take functions that return DOM nodes.
+- Fixed: `pause` and `unpause` cannot accidentally add extra event listeners.
 
 ## 2.2.0